+++ app/src/main/AndroidManifest.xml
@@ -25,7 +25,7 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.ARKDrop">
         <activity
-            android:name=".presentation.qrcodescanner.QRCodeScannerActivity"
+            android:name=".presentation.feature.qrcodescanner.QRCodeScannerActivity"
             android:exported="false"
             android:theme="@style/Theme.AppCompat" />
         <activity
