+++ app/src/main/java/com/arkbuilders/arkdrop/presentation/feature/transferprogress/TransferProgressScreen.kt
@@ -1,4 +1,4 @@
-package com.arkbuilders.arkdrop.presentation.transferprogress
+package com.arkbuilders.arkdrop.presentation.feature.transferprogress
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
@@ -17,7 +17,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.navigation.NavController
 import androidx.navigation.compose.rememberNavController
-import com.arkbuilders.arkdrop.presentation.transferprogress.composables.TransferParticipantHeader
+import com.arkbuilders.arkdrop.presentation.feature.transferprogress.composables.TransferParticipantHeader
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
