+++ app/src/main/java/com/arkbuilders/arkdrop/presentation/navigation/BottomTabBar.kt
@@ -1,4 +1,4 @@
-package com.arkbuilders.arkdrop.navigation
+package com.arkbuilders.arkdrop.presentation.navigation
 
 import androidx.compose.animation.animateColorAsState
 import androidx.compose.material3.Icon
