+++ app/src/main/java/com/arkbuilders/arkdrop/presentation/navigation/NavRegister.kt
@@ -1,11 +1,11 @@
-package com.arkbuilders.arkdrop.navigation
+package com.arkbuilders.arkdrop.presentation.navigation
 
 import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
 import com.arkbuilders.arkdrop.Greeting
-import com.arkbuilders.arkdrop.presentation.filestransfers.FilesTransferScreen
-import com.arkbuilders.arkdrop.presentation.transferconfirmation.TransferConfirmation
+import com.arkbuilders.arkdrop.presentation.feature.filestransfers.FilesTransferScreen
+import com.arkbuilders.arkdrop.presentation.feature.transferconfirmation.TransferConfirmation
 
 fun NavGraphBuilder.navRegistration(navController: NavController) {
     composable(TransfersDestination.route) {
