+++ app/src/main/res/layout/activity_qrcode_scanner.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.qrcodescanner.QRCodeScannerActivity">
+    tools:context=".presentation.feature.qrcodescanner.QRCodeScannerActivity">
 
     <androidx.camera.view.PreviewView
         android:id="@+id/viewFinder"
